requirements.txt
pylibscrypt
protobuf
pycrypto